# api/mobile-openapi.yaml
openapi: 3.1.0
info:
  title: FamilyChat Mobile API
  version: 0.1.0
  description: |
    Content encryption protocol fc1: clients send X-Content-Encryption: fc1 when
    reading chats/messages/files. Missing negotiation or plaintext writes return
    HTTP 426 encryption_required. Obtain the common server key using the authenticated
    GET /encryption/key over HTTPS. This is shared-key encryption, not E2EE.
    Text is base64 of an FCENC001 envelope; file and preview bodies are binary envelopes.
    HPKE suite is X25519/HKDF-SHA256/AES-256-GCM (RFC 9180 base mode).
    See README.md for framing, contexts, migration and backup requirements.
    Контракт действующего Go-сервера для мобильных клиентов. Базовый путь — /api/v1.
    Все операции, кроме регистрации, входа, чтения аватара и WebSocket upgrade, требуют
    Authorization: Bearer <token>. Самостоятельная регистрация доступна только
    когда её включил администратор; телефон, почта и приглашение не нужны.

    Вложения и превью всегда скачиваются через API с Bearer-заголовком.
    До отправки сообщения файл видит только загрузивший его участник чата;
    после отправки — текущие участники. Удаление участника или отзыв сессии
    блокируют последующие запросы. Аватар — отдельное изображение JPEG/PNG:
    текущий аватар выдаётся через публичный GET /users/{userID}/avatar без
    Bearer; это не открывает доступа к вложениям чата. Прямых S3 URL нет.

...
    Для мобильного клиента после разрыва соединения REST-история остаётся
...
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Непрозрачный токен из POST /auth/login; передавать в заголовке, не в URL.
  schemas:
...
    UUID:
      type: string
      format: uuid
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Машиночитаемый код ошибки, например not_found или busy.
...
    Chat:
      type: object
      required: [id, kind, title, role, unread_count]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        kind: {type: string, enum: [direct, group]}
        title:
          type: string
          description: В личном чате это актуальное отображаемое имя собеседника или его логин, если имя не задано.
        role: {type: string, enum: [owner, admin, member]}
        avatar_url: {type: string, description: Публичный аватар собеседника в личном чате; отсутствует для группы и профиля без аватара.}
        unread_count: {type: integer, format: int64, minimum: 0, description: Число сообщений других участников после последней отмеченной позиции прочтения.}
        last_message_ciphertext: {type: string, description: Base64 full encrypted last message. Decrypt locally to create a preview.}
        last_message_sender_id: {$ref: '#/components/schemas/UUID'}
        last_message_client_id: {$ref: '#/components/schemas/UUID'}
        last_message_preview: {type: string, description: Заглушка для зашифрованного сообщения; клиент заменяет её расшифрованным last_message_ciphertext. Отсутствует в пустом чате.}
        last_message_at: {type: string, format: date-time, description: Время последнего сообщения; отсутствует в пустом чате.}
        last_message_sender: {type: string, description: Отображаемое имя отправителя последнего сообщения или логин; используется для превью группы.}
        presence:
          $ref: '#/components/schemas/UserPresence'
          description: Присутствие собеседника; только для личного чата. Онлайн означает хотя бы одно активное WebSocket-подключение.
    UserPresence:
      type: object
      required: [online, last_seen_at]
      properties:
        online: {type: boolean}
        last_seen_at:
          type: [string, 'null']
          format: date-time
          description: Последняя подтверждённая активность подключения. null, если подключений ещё не было. Сохраняется между перезапусками.
...
    GroupInput:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title: {type: string, minLength: 1, maxLength: 100}
...
  parameters:
...
    Limit:
      name: limit
      in: query
      description: По умолчанию 50; значения выше 100 ограничиваются сервером до 100.
      schema: {type: integer, minimum: 1, maximum: 100, default: 50}
...
  responses:
    BadRequest:
      description: Неверный ввод или запрос.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Unauthorized:
      description: Нет действующего Bearer-токена. Middleware возвращает JSON-текст с Content-Type text/plain.
      content:
        text/plain:
          schema: {type: string}
          example: '{"error":"unauthorized"}'
...
paths:
...
  /chats:
    get:
      tags: [Chats]
      operationId: listChats
      summary: Последние созданные чаты текущего пользователя
      description: Пока доступен только limit; курсор списка чатов сервер не поддерживает.
      parameters:
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: Список чатов.
          content:
            application/json:
              schema:
                type: object
                required: [chats]
                properties:
                  chats:
                    type: array
                    items: {$ref: '#/components/schemas/Chat'}
        '401': {$ref: '#/components/responses/Unauthorized'}
    post:
      tags: [Chats]
      operationId: createGroupChat
      summary: Создать групповой чат
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/GroupInput'}
      responses:
        '201':
          description: Созданная группа; автор — owner.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Chat'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
...